@@ -1,13 +1,15 @@
-/* eslint no-alert: 0, jsx-a11y/accessible-emoji: 0  */
-
 import React, { Component } from 'react';
+import { Asset, AppLoading } from 'expo';
 import { View, StyleSheet, Linking } from 'react-native';
 
 import { GiftedChat } from 'react-native-gifted-chat';
+import Sentry from 'sentry-expo';
 
 import messagesData from './data';
 import NavBar from './NavBar';
 import CustomView from './CustomView';
+
+Sentry.config('https://2a164b1e89424a5aafc186da811308cb@sentry.io/276804').install();
 
 const styles = StyleSheet.create({
   container: { flex: 1 },
@@ -20,18 +22,21 @@ export default class App extends Component {
 
   constructor(props) {
     super(props);
+
     this.state = {
       messages: [],
       step: 0,
+      appIsReady: false,
     };
 
     this.onSend = this.onSend.bind(this);
     this.parsePatterns = this.parsePatterns.bind(this);
   }
 
-  componentWillMount() {
+  async componentWillMount() {
     // init with only system messages
-    this.setState({ messages: messagesData.filter((message) => message.system) });
+    await Asset.fromModule(require('./assets/avatar.png')).downloadAsync();
+    this.setState({ messages: messagesData.filter((message) => message.system), appIsReady: true });
   }
 
   onSend(messages = []) {
@@ -40,7 +45,7 @@ export default class App extends Component {
       messages: GiftedChat.append(previousState.messages, [{ ...messages[0], sent: true, received: true }]),
       step,
     }));
-    setTimeout(() => this.botSend(step), 1500 + Math.round(Math.random() * 1000));
+    setTimeout(() => this.botSend(step), 1200 + Math.round(Math.random() * 1000));
   }
 
   botSend(step = 0) {
@@ -59,19 +64,23 @@ export default class App extends Component {
     return [
       {
         pattern: /#(\w+)/,
-        style: { ...linkStyle, color: 'orange' },
+        style: { ...linkStyle, color: 'darkorange' },
         onPress: () => Linking.openURL('http://gifted.chat'),
       },
     ];
   }
   render() {
+    if (!this.state.appIsReady) {
+      return <AppLoading />;
+    }
     return (
-      <View style={styles.container}>
+      <View style={styles.container} accessible accessibilityLabel="main" testID="main">
         <NavBar />
         <GiftedChat
           messages={this.state.messages}
           onSend={this.onSend}
           renderCustomView={CustomView}
+          keyboardShouldPersistTaps="never"
           user={{
             _id: 1,
           }}
