@@ -1,12 +1,38 @@
-import React, { Component } from 'react';
-import { View, Platform } from 'react-native';
+import React from 'react';
+import { Platform } from 'react-native';
 import PropTypes from 'prop-types';
 import { GiftedChat } from 'react-native-gifted-chat';
 import emojiUtils from 'emoji-utils';
 
 import SlackMessage from './SlackMessage';
 
-class App extends Component {
+export default class App extends React.Component {
+  state = {
+    messages: [],
+  }
+
+  componentWillMount() {
+    this.setState({
+      messages: [
+        {
+          _id: 1,
+          text: 'Hello developer',
+          createdAt: new Date(),
+          user: {
+            _id: 2,
+            name: 'React Native',
+            avatar: 'https://placeimg.com/140/140/any',
+          },
+        },
+      ],
+    })
+  }
+
+  onSend(messages = []) {
+    this.setState(previousState => ({
+      messages: GiftedChat.append(previousState.messages, messages),
+    }))
+  }
 
   renderMessage(props) {
     const { currentMessage: { text: currText } } = props;
@@ -30,12 +56,14 @@ class App extends Component {
   render() {
     return (
       <GiftedChat
-        messages={[]}
+        messages={this.state.messages}
+        onSend={messages => this.onSend(messages)}
+        user={{
+          _id: 1,
+        }}
         renderMessage={this.renderMessage}
       />
     );
   }
 
 }
-
-export default App;
