@@ -13,25 +13,20 @@ export default class Actions extends React.Component {
   }
 
   onActionsPress() {
-    const options = Object.keys(this.props.options);
-    const cancelButtonIndex = Object.keys(this.props.options).length - 1;
+    const { options } = this.props;
+    const optionKeys = Object.keys(options);
+    const cancelButtonIndex = optionKeys.indexOf('Cancel');
     this.context.actionSheet().showActionSheetWithOptions(
       {
-        options,
+        options: optionKeys,
         cancelButtonIndex,
         tintColor: this.props.optionTintColor,
       },
-      function handle(buttonIndex) {
-        let i = 0;
-        Object.keys(this.props.options).forEach(function launch(key) {
-          if (this.props.options[key]) {
-            if (buttonIndex === i) {
-              this.props.options[key](this.props);
-              return;
-            }
-            i += 1;
-          }
-        });
+      (buttonIndex) => {
+        const key = optionKeys[buttonIndex];
+        if (key) {
+          options[key](this.props);
+        }
       },
     );
   }
