@@ -17,6 +17,10 @@ export default class MessageText extends React.Component {
     this.onEmailPress = this.onEmailPress.bind(this);
   }
 
+  shouldComponentUpdate(nextProps) {
+    return this.props.currentMessage.text !== nextProps.currentMessage.text;
+  }
+
   onUrlPress(url) {
     // When someone sends a message that includes a website address beginning with "www." (omitting the scheme),
     // react-native-parsed-text recognizes it as a valid url, but Linking fails to open due to the missing scheme.
@@ -62,17 +66,9 @@ export default class MessageText extends React.Component {
   }
 
   render() {
-    const linkStyle = StyleSheet.flatten([
-      styles[this.props.position].link,
-      this.props.linkStyle[this.props.position],
-    ]);
+    const linkStyle = StyleSheet.flatten([styles[this.props.position].link, this.props.linkStyle[this.props.position]]);
     return (
-      <View
-        style={[
-          styles[this.props.position].container,
-          this.props.containerStyle[this.props.position],
-        ]}
-      >
+      <View style={[styles[this.props.position].container, this.props.containerStyle[this.props.position]]}>
         <ParsedText
           style={[
             styles[this.props.position].text,
