@@ -5,10 +5,13 @@ import React from 'react';
 import { StyleSheet, Text, TouchableOpacity, View, ViewPropTypes } from 'react-native';
 import Color from './Color';
 
-export default function Send({ text, containerStyle, onSend, children, textStyle, label }) {
-  if (text.trim().length > 0) {
+export default function Send({ text, containerStyle, onSend, children, textStyle, label, alwaysShowSend }) {
+  if (alwaysShowSend || text.trim().length > 0) {
     return (
       <TouchableOpacity
+        testID="send"
+        accessible
+        accessibilityLabel="send"
         style={[styles.container, containerStyle]}
         onPress={() => {
           onSend({ text: text.trim() }, true);
@@ -45,6 +48,7 @@ Send.defaultProps = {
   containerStyle: {},
   textStyle: {},
   children: null,
+  alwaysShowSend: false,
 };
 
 Send.propTypes = {
@@ -54,4 +58,5 @@ Send.propTypes = {
   containerStyle: ViewPropTypes.style,
   textStyle: Text.propTypes.style,
   children: PropTypes.element,
+  alwaysShowSend: PropTypes.bool,
 };
