@@ -9,7 +9,7 @@ import moment from 'moment';
 import Color from './Color';
 import { TIME_FORMAT } from './Constant';
 
-export default function Time({ position, containerStyle, currentMessage, timeFormat }, context) {
+export default function Time({ position, containerStyle, currentMessage, timeFormat, textStyle }, context) {
   return (
     <View style={[styles[position].container, containerStyle[position]]}>
       <Text style={[styles[position].text, textStyle[position]]}>
